% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R7multifile-package.R
\docType{package}
\name{R7multifile-package}
\alias{R7multifile}
\alias{R7multifile-package}
\title{R7multifile: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KatherineCox/R7multifile}
  \item Report bugs at \url{https://github.com/KatherineCox/R7multifile/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
